+++ src/errors/InvalidAuthorizationHeaderError.ts
@@ -1,10 +1,11 @@
-import ApplicationError from './ApplicationError';
+import { ApplicationError } from '@piggly/ddd-toolkit';
+import { crc32 } from 'crc';
 
 /**
  * @file This error should be thrown when authorization header is invalid.
  * @copyright Piggly Lab 2023
  */
-export default class InvalidAuthorizationHeaderError extends ApplicationError {
+export class InvalidAuthorizationHeaderError extends ApplicationError {
 	/**
 	 * Create a new error.
 	 *
@@ -18,7 +19,7 @@ export default class InvalidAuthorizationHeaderError extends ApplicationError {
 	constructor(message?: string, hint?: string, extra?: Record<string, any>) {
 		super(
 			'InvalidAuthorizationHeaderError',
-			51,
+			crc32('InvalidAuthorizationHeaderError'),
 			message || 'Invalid authorization header.',
 			401,
 			hint || 'The `Authorization` header must be of `Bearer` type.',