+++ src/errors/RequestNotFoundError.ts
@@ -1,10 +1,11 @@
-import ApplicationError from './ApplicationError';
+import { ApplicationError } from '@piggly/ddd-toolkit';
+import { crc32 } from 'crc';
 
 /**
  * @file This error should be thrown when request is not found.
  * @copyright Piggly Lab 2023
  */
-export default class RequestNotFoundError extends ApplicationError {
+export class RequestNotFoundError extends ApplicationError {
 	/**
 	 * Create a new error.
 	 *
@@ -18,7 +19,7 @@ export default class RequestNotFoundError extends ApplicationError {
 	constructor(message?: string, hint?: string, extra?: Record<string, any>) {
 		super(
 			'RequestNotFoundError',
-			56,
+			crc32('RequestNotFoundError'),
 			message || 'Cannot access requested URL.',
 			404,
 			hint || 'You must check the URL or the request parameters.',