+++ src/errors/index.ts
@@ -1,11 +1,8 @@
-export { default as RuntimeError } from './RuntimeError';
-export { default as ApplicationError } from './ApplicationError';
-
-export { default as RequestNotFoundError } from './RequestNotFoundError';
-export { default as RequestServerError } from './RequestServerError';
-export { default as InvalidRequestBodyError } from './InvalidRequestBodyError';
-export { default as InvalidRequestQueryError } from './InvalidRequestQueryError';
-export { default as MissingAuthorizationHeaderError } from './MissingAuthorizationHeaderError';
-export { default as InvalidAuthorizationHeaderError } from './InvalidAuthorizationHeaderError';
-export { default as ForbiddenError } from './ForbiddenError';
-export { default as UnauthorizedError } from './UnauthorizedError';
+export { RequestNotFoundError } from './RequestNotFoundError';
+export { RequestServerError } from './RequestServerError';
+export { InvalidRequestBodyError } from './InvalidRequestBodyError';
+export { InvalidRequestQueryError } from './InvalidRequestQueryError';
+export { MissingAuthorizationHeaderError } from './MissingAuthorizationHeaderError';
+export { InvalidAuthorizationHeaderError } from './InvalidAuthorizationHeaderError';
+export { ForbiddenError } from './ForbiddenError';
+export { UnauthorizedError } from './UnauthorizedError';