+++ src/index.ts
@@ -1,8 +1,66 @@
-export * from './types';
-export * from './errors';
-export * from './helpers';
-export * from './utils';
-export * from './services';
-export * from './core';
-export * from './hooks';
-export * from './www';
+export type {
+	EnvironmentType,
+	DefaultEnvironment,
+	EnvironmentAccessTokenOptions,
+	EnvironmentMysqlOptions,
+	FastifyServer,
+	FastifyModifierCallable,
+	FastifyAppliable,
+	ApiServerOptions,
+	ApiServerInterface,
+	HttpServerInterface,
+	RuleInterface,
+	JWTServiceOptions,
+	JWTServiceInterface,
+	AccessTokenServiceOptions,
+	AccessTokenServiceErrors,
+	SchemaHandler,
+	PaginationMetaProps,
+	PaginationMetaJSON,
+	INVALID_ACCESS_TOKEN_EVENT,
+} from './types';
+
+export {
+	RequestNotFoundError,
+	RequestServerError,
+	InvalidRequestBodyError,
+	InvalidRequestQueryError,
+	MissingAuthorizationHeaderError,
+	InvalidAuthorizationHeaderError,
+	ForbiddenError,
+	UnauthorizedError,
+} from './errors';
+
+export { Logger } from './helpers';
+
+export {
+	getBearerToken,
+	getBasicToken,
+	mountURL,
+	getIp,
+	getOrigin,
+	PaginationMeta,
+} from './utils';
+
+export {
+	JWTAccessTokenService,
+	JWTService,
+	JWTEdDSAService,
+	HealthCheckService,
+} from './services';
+
+export { BaseController } from './core';
+
+export { SyncErrorOnDiskHandler } from './handlers';
+
+export { AuditRequestLogger } from './hooks';
+
+export {
+	AbstractServer,
+	HttpInsecureServer,
+	HttpSecureServer,
+	Http2InsecureServer,
+	Http2SecureServer,
+	FastifyModifiers,
+	HttpServer,
+} from './www';