+++ src/utils/index.ts
@@ -1,3 +1,9 @@
-export * from './globals';
-export * from './parsers';
-export { default as PaginationMeta } from './PaginationMeta';
+export {
+	getBearerToken,
+	getBasicToken,
+	mountURL,
+	getIp,
+	getOrigin,
+} from './globals';
+
+export { PaginationMeta } from './PaginationMeta';