+++ src/utils/parsers/index.ts
@@ -1,1 +1,0 @@
-export { default as DateParser } from './DateParser';